bugfix: set vars right away when sourcing tests/default
sometimes you need to refer to a variable (for example
num_procs=$num_procs_down), allowing you to source some common logic
which just uses $num_procs (such as a test/generic_*), and having
to override the test_setvars callback would be very cumbersome.

so we leverage the fact that the needed vars ($src, $test, $project)
are set when tests/default.sh is sourced, so it can in turn set the
variables immediately, so that the actual testcase file immediately
has acces to them.

the fact that this went unnoticed without this commit, is because
vars will remain set after the first test case, and the first
testcase was trivial enough to not require this.